requirements.txt
PyYAML==6.0.1
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
networkx==3.2.1
python-multipart==0.0.6
aiofiles==23.2.1
jsonschema==4.20.0
pydantic==2.5.0
python-multipart==0.0.6
